+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+### 0.2.0-beta
+
+- New command to modify a Lightning Web Component
+- Fixed repeated request issue delaying the code generation
+
 ### 0.1.0-beta
 
 - New command to generate a Lightning Web Component